Expose code generation routines as non-internal `code-gen` module
The code generation routines are moved from `tool-base` to a new
`code-gen` module and are now mostly non-`Internal`. This way they
can be reused by implementors of protoc code generators and others.
This is especially relevant for column and field generation as
Spine routines here perform the logic which would be common for most
implementations.